Create dedicated layout for News pages
Note: This is different from the Post layout. That layout is for
individual posts, while this layout is for pages that list multiple
posts (/news/ and /news/categories/). It makes more sense to have a
dedicated layout for this rater than relying on an ad-hoc exclusion
in the Post layout.