Differential spell checking
This script does not attempt to find all misspellings, but rather show
when new words not previously known are introduced. This mean it's
useful in a continuous integration pipeline where we don't necessarily
want each contributor to have to be responsible for the bad spelling of
whoever came before them, but rather at least not make things worse.
The existing spelling can then be incrementally cleaned up.

This approach also means we don't need to maintain a local dictionary to
avoid false-positives, because existing "misspelled words" ("AppVM",
"marmarek", etc.) will be ignored.

In adition, effort is made to only check prose, not code. All the actual
parsing logic is handled by pandoc, and we just filter and transform a
JSON AST that it produces.

We also operate on the fully built HTML site rather than the source
markdown, meaning we see all spellings as they exist post-templating,
including all values from _data/*.yaml and such already filled in.

This was written primarily for use in a CI hook, but is of course also
useful locally. The default invocation checks the net difference between
origin/master and your local HEAD. You can also set SPELL_CHECK_CACHE to
a directory in which to keep results of previous runs, which speeds up
re-checking incremental changes.

Introduced dependencies are: hunspell (preferred) or aspell, and jq.

Fixes: https://github.com/QubesOS/qubes-issues/issues/2725