travis: Run spell check on PRs
We only do this on PRs because we are comparing the currently
unrecognized spellings to the previously unrecognized spellings.
We can only do this comparison if we have something to compare against.

Run all the checks all the time and report if any failed rather than
stopping at first failure. This may reduce the number of iterations
required to fix all problems with a given PR.