Add XSA-273 with TBA status
This is an attempt to stem the tide of questions we're receiving about
XSA-273 on a daily basis. Hitherto, our policy has been to add an XSA
to the Tracker only with a determinate "yes" or "no" answer to the
question of whether the security of Qubes is affected by that XSA.
However, we did not account for the possibility that all members of the
Qubes Security Team might simultaneously be unavailable for an extended
period of time while a critical XSA is released, and hence that we might
be unable to issue a determinate answer in a timely manner.

Related commit: https://github.com/QubesOS/qubes-doc/commit/af494898eacdd1bcb7c2cbbf7a03d62f6b9eecb3