Update community icon class name
Not all of these icons links are for social media platforms. In
particular, there is an internal link to our own support page, as well
as a link to the Qubes forum. This change is intended to help visitors
who use a browser extension to block social media icons but who may
still want to see the aforementioned non-social-media icons.