Make doc sidebar sticky
Creates a "sidebar" class as the sticky target and moves the "advanced"
and "external" doc alert warnings into the main doc content pane to
avoid overlapping sticky elements near the top of the page.

The translation warning is fine in its current location. Since it's
above everything else on the page and isn't sticky, it doesn't overlap
with any sticky elements. However, if we ever make it sticky in the
future, we'll probably need to adjust its position accordingly.

Note that sticky behavior on small viewports (such as mobile devices)
was not observed to be working before this commit.