Fix sticky ToC when sidebar is not shown
When the viewport is small enough to hide the sidebar, the "stickiness"
still applied to the table of contents on that page, causing it overlap
the text illegibly.

This commit restricts sidebar stickiness so that it applies only when
the viewport is large enough for the sidebar to be shown.

An alternative way to address this is to make the background of the
table of contents opaque so that the text is still legible, but this
would arguably take up too much screen real estate, requiring the user
to press the "Contents" button in order to collapse the table of
contents on each affected page.

Note that alert stickiness is not affected. It stil makes sense for
alerts to exhibit their current stickiness behavior even when the
viewport is small enough for the sidebar to be hiden.