Remove Internet Archive entry
It's not the same type of "download mirror" as the other entries in this
list and usually doesn't have the latest ISO. This is understandable
given the Internet Archive's function and purpose, but makes it
unsuitable as an everyday ISO download source for users. It would make
more sense to archive Qubes artifacts here for historical interest
rather than attempt to use it as a software delivery mechanism,
especially when we have so many other download mirrors designed for and
already performing that exact job.