Update doc-index.yml
Moved `Glossary` to "Introduction," and `CLI Tools` to "Advanced Topics."
#3 and #4 in this comment: https://github.com/QubesOS/qubes-issues/issues/6756#issuecomment-873456424

TL;DR, this is where people are more likely to find them, than per organizational structures more "correct" to library sciences or encyclopedic knowledge management standards. 

"Introduction" is really the section for the 5 things people need to absolutely not miss, before bothering everyone else with a question. "CLI Tools" is also findable in "Advanced Topics," I'd speculate—moreso than in a "Reference" section. 

Not keen to die on this hill though, so am happy to discuss if there's strong pushback.