Add glossary button to sidebar; update ToC button text
Thanks to @deeplow for suggesting the addition of the glossary button
in https://github.com/QubesOS/qubesos.github.io/pull/181#issuecomment-874099956.

Also removes "Go to the" from the "Table of Contents" button. Even
though all the other buttons are imperative statements, it looks
weird with this one for some reason. Since this button is colored
differently from the others, hopefully this'll be ok.