Use standard site layout for Team page
The added maintenance burden of having a completely separate layout and
a completely separate page heading for only a single page just to add a
few links to the top (that are probably ignored anyway) is not worth it.
If abuse of the Team page proves to be a problem (e.g., sending help and
support questions to the email addresses on this page rather than to the
proper venues), then we should pursue a strategy that is more likely to
be effective. A small link among several others at the top is not likely
to be effective. Moreover, this separate layout made the heading height
on this page inconsistent with all the others, which looked somewhat
unprofessional.