Change doc index h1 to "Table of contents"
"Qubes documentation" and varaints are redundant with the large
"Documentation" page heading above, and this more clearly states the
nature of this page.

A minor problem introduced by this change is that it might be confusing
to have both "Page contents" in the sidebar as well as "Table of
contents" here, but attentive readers should be able to tell that "Page
contents" refers to the contents of this page, where as "Table of
contents" refers to the contents of the documentation as a whole. At any
rate, this still seems like the lesser of two evils compared to "Qubes
documentation" (and variants).