Correct percentage for administrative fees
I rechecked the Open Collective source & saw that the fees are 5% via
crowdfunding or 8% via bank transfer/check. There are reductions for
the latter, but only if at least $500k is raised (annually, I think).

So, the "final" percentage for admin fees would be a weighted average
between 5% and 8%.

The fees may have been 14% at some point in the past, but that is not
the administrative fee rate today.