Add 'rootless' Make target
Creates a new Make target 'rootless' that seeks to avoid the need for
routine use of elevated privileges.

Initial installation of podman, git, and other tools into the
template as root is necessary. Running the development loop
as an unprivileged user in the app qube avoids unneeded access.

An additional benefit is seen with standard app qubes. Running the
process with 'sudo' will write files such as the container images to
the volatile parts of storage. The user would need to re-download the
container images after each qube restart.