Reference software versions that build the site properly
The instructions for using the jekyll/jekyll:pages podman image were no
longer working for me. This is most likely due to the image being a
rolling release rather than a fixed release and in the meantime changes
got introduced that broke building the site locally.

The compose.yml file got a change introduced so from now on the images
is always of a fixed version. Furthermore, README.md has been updates so
it references the environment used for testing as of today.