Add 'webrick' gem if website is being built locally
Webrick used to be part of the Ruby standard library in version 2. It
was removed in Ruby version 3.

Jekyll lists webrick as a dependency. It does not install it by
default, however. Jekyll can be used to simply build a website and
then a different program can serve it. Webrick is only needed if
Jekyll will need to serve files. The command run in the podman
container is 'jekyll serve' and will require webrick.

As webrick is not needed if Jekyll is only building the website, the
gem statement is conditional upon an environment variable set by the
podman command.