Have podman remove the container after 'jekyll serve' finishes
The '--rm' flag removes the container when it exits.

If the intention of using containers is to create a more consistent
development environment, explicitly removing the container between
runs will help prevent inconsistencies due to a container's internal
state.

The image upon which the container is based is untouched by this flag.